OGS-482
